PAL out SocketAsyncEventArgs and Socket.SendMultiple.
