+++ src/System.Net.Sockets/src/System/Net/Sockets/_OverlappedAsyncResult.Unix.cs
@@ -7,14 +7,21 @@ namespace System.Net.Sockets
 {
     internal partial class OverlappedAsyncResult : BaseOverlappedAsyncResult
     {
-        public new void CompletionCallback(int numBytes, byte[] socketAddress, int socketAddressLen, SocketError errorCode)
+        int _socketAddressSize;
+
+        internal int GetSocketAddressSize()
+        {
+            return _socketAddressSize;
+        }
+
+        public new void CompletionCallback(int numBytes, byte[] socketAddress, int socketAddressSize, SocketError errorCode)
         {
             ErrorCode = (int)errorCode;
 
             if (_socketAddress != null)
             {
                 Debug.Assert(socketAddress == null || _socketAddress.Buffer == socketAddress);
-                _socketAddress.InternalSize = socketAddressLen;
+                _socketAddressSize = socketAddressSize;
             }
 
             InvokeCallback(PostCompletion(numBytes));
